+++ app/schemas/transaction.py
@@ -1,6 +1,8 @@
 from typing import Optional
 
 from pydantic import BaseModel
+from decimal import Decimal
+
 
 
 class TransactionIn(BaseModel):
@@ -12,19 +14,19 @@ class TransactionOut(BaseModel):
     id: int
     stock: str
     volume: int
-    price: Optional[float]
-    purchase_price: Optional[float] = 0
+    price: Optional[Decimal]
+    purchase_price: Optional[Decimal] = 0
     purchase_date: str
-    current_price: Optional[float] = None
-    gain: Optional[float] = 0
+    current_price: Optional[Decimal] = None
+    gain: Optional[Decimal] = 0
 
 
 class TransactionOutWithTotal(BaseModel):
-    total_gain: Optional[float] = 0
-    total_value: Optional[float] = 0
+    total_gain: Optional[Decimal] = 0
+    total_value: Optional[Decimal] = 0
     transactions: list[TransactionOut]
 
 
 class CreateTransactionResponse(BaseModel):
     success: bool
-    balance: float
+    balance: Decimal
